Moves pagination logic to client
