+++ integrations/gitlab-v2/clients/gitlab_client.py
@@ -29,66 +29,10 @@ class GitLabClient:
             if projects_batch:
                 yield projects_batch
 
-            async for updated_batch in self._process_nested_fields(
+            async for updated_batch in self.graphql._process_nested_fields(
                 projects_batch, field_iterators
             ):
                 yield updated_batch
-
-    async def _process_nested_fields(
-        self,
-        projects: list[dict[str, Any]],
-        field_iterators: list[AsyncIterator[tuple[str, list[dict[str, Any]]]]],
-    ) -> AsyncIterator[list[dict[str, Any]]]:
-        """Process nested fields for a batch of projects, yielding after meaningful updates."""
-
-        project_field_nodes: list[dict[str, list[dict[str, Any]]]] = [
-            {} for _ in projects
-        ]
-        active_data = list(zip(projects, field_iterators, project_field_nodes))
-
-        while active_data:
-            updated = False
-            next_active = []
-            # Parallelize awaits for all iterators
-            tasks = [anext(field_iter) for _, field_iter, _ in active_data]
-            results = await asyncio.gather(*tasks, return_exceptions=True)
-
-            for (project, field_iter, field_nodes), result in zip(active_data, results):
-                if isinstance(result, StopAsyncIteration):
-                    logger.debug(
-                        f"Field iterator exhausted for project {project.get('id', 'unknown')}"
-                    )
-                    continue
-
-                if isinstance(result, Exception):
-                    logger.error(f"Error in field iterator: {result}")
-                    continue
-
-                field_name, nodes = cast(tuple[str, list[dict[str, Any]]], result)
-                if nodes:
-                    if field_name not in project:
-                        project[field_name] = {"nodes": []}
-                    elif (
-                        not isinstance(project[field_name], dict)
-                        or "nodes" not in project[field_name]
-                    ):
-                        project[field_name] = (
-                            {"nodes": project[field_name]}
-                            if project[field_name]
-                            else {"nodes": []}
-                        )
-
-                    field_nodes.setdefault(field_name, []).extend(nodes)
-                    project[field_name]["nodes"] = field_nodes[field_name]
-                    updated = True
-
-                next_active.append((project, field_iter, field_nodes))
-
-            active_data = next_active
-            if updated:
-                logger.info(f"Yielding batch with {len(projects)} projects")
-                yield [deepcopy(project) for project, _, _ in active_data]
-
     async def get_groups(self) -> AsyncIterator[list[dict[str, Any]]]:
         """Fetch all groups accessible to the user."""
         async for batch in self.rest.get_resource(
